[refactor] 로그인/회원가입 페이지 배경 공유를 위한 작업
